[fix] : (TodoList.tsx, Calendar.tsx) 컴포넌트 로직 분리 1차 정리 세부 내용1 : Calendar 컴포넌트를 FullCalendar 관련 로직으로만 구성 세부 내용2 : TodoList 컴포넌트를 오늘의 투두와 완료된 투두를 표시하는 로직으로 구성 세부 내용3 : (todo-list.page.tsx) 상위 페이지에서 투두 페칭
