+++ src/app/(main)/todo-list/page.tsx
@@ -1,7 +1,16 @@
+import { fetchTodos } from "@/utils/todoApi";
 import Calendar from "../_components/Calendar";
+import TodoList from "./_components/TodoList";
 
-const TodoListPage = () => {
-  return <Calendar />;
+const Page = async () => {
+  const todos = await fetchTodos();
+
+  return (
+    <div>
+      <Calendar todos={todos} />
+      <TodoList todos={todos} />
+    </div>
+  );
 };
 
-export default TodoListPage;
+export default Page;
